add election controlled service endpoints
Manually setting the endpoints for a service ensures that only the
elected pods receive traffic. This works around issues where controlling
traffic via pod readiness works suboptimal (i.e. rolling deployments).

In rolling deployments, readiness of a pod controls the further roll-out
of new versions. This means using readiness to control the service proxy
would lead to a deadlock: the new pods would never become "ready" as an old
pod still holds the lease, but the old pod is never terminated, as the rolling
update only continues if the new pods are ready.